+++ scripts/git-subtree/inbox/add-squash.ps1
@@ -1,0 +1,5 @@
+git subtree add `
+    -P workspace/plugins/inbox `
+    https://github.com/Zuoqiu-Yingyi/siyuan-plugin-inbox.git `
+    dev `
+    --squash