+++ scripts/git-subtree/inbox/pull.ps1
@@ -1,0 +1,4 @@
+git subtree pull `
+    -P workspace/plugins/inbox `
+    https://github.com/Zuoqiu-Yingyi/siyuan-plugin-inbox.git `
+    dev