+++ scripts/git-subtree/inbox/push.ps1
@@ -1,0 +1,4 @@
+git subtree push `
+    -P workspace/plugins/inbox `
+    https://github.com/Zuoqiu-Yingyi/siyuan-plugin-inbox.git `
+    dev